Move vendor functions into clash_utils
