+++ src/ifcgeom_schema_agnostic/clash_utils.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include <gp_Vec.hxx>
+#include <array>
+
+struct ray {
+    float origin[3];
+    float dir[3];
+    float dir_inv[3];
+};
+
+struct box {
+    float corners[2][3];
+};
+
+bool is_intersect_ray_box(const struct ray *ray, const struct box *box);
+
+bool intersectRayTriangle(	const gp_Vec& orig, const gp_Vec& dir, 
+                                                const gp_Vec& vert0, const gp_Vec& vert1, const gp_Vec& vert2, 
+                                                Standard_Real& at, Standard_Real& au, Standard_Real& av,
+                                                bool cull, float enlarge=0.0f);
+
+void edgeEdgeDist(gp_Vec& x, gp_Vec& y,				// closest points
+                 const gp_Vec& p, const gp_Vec& a,	// seg 1 origin, vector
+                 const gp_Vec& q, const gp_Vec& b);	// seg 2 origin, vector
+
+float distanceTriangleTriangleSquared(gp_Vec& cp, gp_Vec& cq, const std::array<gp_Vec, 3> p, const std::array<gp_Vec, 3> q);
